remove generate-partial-fasta as subcommand and add it is an alternate top-level entrypoint (main method)
