+++ src/main/scala/org/hammerlab/guacamole/Guacamole.scala
@@ -32,8 +32,7 @@ import org.hammerlab.guacamole.commands.jointcaller.SomaticJoint
 object Guacamole extends Logging {
 
   /**
-   * Commands (i.e. variant caller implementations) that are part of Guacamole. If you add a new variant caller, update
-   * this list.
+   * Commands (e.g. variant caller implementations) that are part of Guacamole. If you add a new command, update this.
    */
   private val commands: Seq[Command[_]] = List(
     GermlineThreshold.Caller,
@@ -42,9 +41,7 @@ object Guacamole extends Logging {
     SomaticStandard.Caller,
     VariantSupport.Caller,
     VAFHistogram.Caller,
-    SomaticJoint.Caller,
-    GeneratePartialFasta.Caller
-  )
+    SomaticJoint.Caller)
 
   private def printUsage() = {
     println("Usage: java ... <command> [other args]\n")
@@ -57,6 +54,7 @@ object Guacamole extends Logging {
 
   /**
    * Entry point into Guacamole.
+   *
    * @param args command line arguments
    */
   def main(args: Array[String]): Unit = {
